fix: クォート境界でトークンが分割される問題を修正
expand_parameter のクォート開始/終了時に push_token_expand を
呼んでいたため l"s" が [l, "s"] の2トークンに分割されていた。
クォート遷移時は buffer に文字を追加するだけに修正。
