+++ Makefile
@@ -129,6 +129,7 @@ SRCS_MAND	:=	src/main.c \
 				src/expand/expand_parameter/store/free_store_expand.c \
 				src/expand/expand_tilde/expand_tilde.c \
 				src/expand/expand_wildcard/resolve_wildcard.c \
+				src/expand/expand_wildcard/match_wildcard.c \
 				src/expand/expand_wildcard/expand_wildcard.c \
 				src/expand/expand_remove_quotes/expand_remove_quotes.c \
 				src/signal/setup_signal_handlers.c \
